+++ docs/changelog.md
@@ -17,10 +17,12 @@ Each release rewinds and fast-forwards your mixtape experience—just like a rea
 - Floating buttons in editor for mobile to:
     - Jump to mixtape tracks
     - Save changes
+- Added button for manual resyncing the music collection library
 
 ### 🐛 Fixed
 
 - Saving existing mixtapes results in error
+- Covers not available for non-admin users
 
 [v0.3.1](https://github.com/mark-me/mixtape-society/compare/v0.3.0...v0.4.0)